sample_tweepy/requirements.txt
jsonpickle==1.2
tweepy==3.8.0
python-dotenv==0.10.3
